added silent parameter to stop debugger from trying to remove cron
- added silent parameter in the /analyze endpoint
- added silent parameter handling in debugger (if siilent is set, debugger doesn't send /delete request)

+++ debugger.php
@@ -2120,8 +2120,10 @@ function login($password)
  */
 function deleteCronAndCache()
 {
-    $cronAPI = new CronAPI();
-    $cronAPI->deleteCron();
+    if (!isset($_GET['silent'])) {  // if silent is set, don't try to delete cron as the removal request is already sent by EasyWP Cron.
+        $cronAPI = new CronAPI();
+        $cronAPI->deleteCron();
+    }
 
     try {
         $cache = new EasyWP_Cache();
